requirements.txt
djangorestframework==3.14.0
requests==2.31.0
django-storages==1.13.2
Django~=4.2.14
django-cors-headers==4.3.1
librosa
numpy
onnxruntime
psutil
transformers
fuzzywuzzy
fuzz